requirements.txt
streamlit
requests
newspaper3k
transformers
torch
lxml[html_clean]
Pillow
ipykernal
